backend/requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
more-itertools==10.5.0
pillow==10.4.0
premailer==3.10.0
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
yagmail==0.15.293
